Several bug corrections and code adjustments. It builds again in Windows
- In AEventManager, AShower and AMisEt are now exported to the library
- Corrected a simple (but fatal) bug in the definitions of AEventObject::eType which made AMELIA crash every time a MissingEt track was bookmarked.
- in ATrackCombination::recalculate() _theta is calculated with a static cast over one of the arguments of atan2. This avoids complains from the compiler about undefined overloaded functions

+++ src/aeventmanager/ATrackCombination.cpp
@@ -145,7 +145,7 @@ void ATrackCombination::recalculate()
   /* Update other kinematic variables */
   _pt=sqrt(Px*Px + Py*Py);
   _phi=atan2(Py,Px);
-  _theta=atan2(_pt,Pz); // I think this uses the right axis
+  _theta=atan2(static_cast<double>(_pt),Pz); // I think this uses the right axis
   _eta=-logf(tan(_theta/2)); //Pseudo-rapidity is not additive, right?
   _et=E*sin(_theta);
 }
